Update patch set 1
I would prefer that you didn't submit this.

I don't understand how this is related to Freerunner. The change looks
like you're disabling a build-time error check (which implies that you
ran into that error), but I'd like to understand why the check is
unnecessary (and, since the change is in an external library, whether it
should be pushed upstream).

Patch-set: 1